backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.2.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
deprecation==2.1.0
easyocr==1.7.2
exceptiongroup==1.3.1
fastapi==0.128.0
filelock==3.19.1
fsspec==2025.10.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.13.0.90
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==11.3.0
pillow-heif==1.1.1
postgrest==1.1.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
realtime==2.7.0
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
six==1.17.0
starlette==0.49.3
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
sympy==1.14.0
tokenizers==0.22.2
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.2
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.39.0
websockets==15.0.1
